services/api/requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
supabase==2.13.0
openai==1.82.0
deepgram-sdk==3.9.0
python-multipart==0.0.20
pydantic-settings==2.8.1
python-pptx==1.0.2
python-docx==1.1.2
PyPDF2==3.0.1
pymupdf==1.25.4
tenacity==9.1.2
httpx==0.28.1
reportlab>=4.0.0
stripe>=8.0.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
google-auth>=2.25.0
